+++ src/modules/auth/components/AuthPasswordChanged/AuthPasswordChanged.tsx
@@ -1,14 +1,14 @@
 import { ReactComponent as CheckCircleOutlined } from '../../../shared/assets/icons/checkCircledOutlined.svg';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import { useCompanyConfig } from '../../../shared/hooks/useCompanyConfig';
-import useRouter from '../../../shared/hooks/useRouter';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
 import useTranslation from '../../../shared/hooks/useTranslation';
 import { AuthButton } from '../AuthButton';
 import { AuthFooter } from '../AuthFooter';
 import { AuthLayoutBase } from '../AuthLayoutBase';
 
 export const AuthPasswordChanged = () => {
-  const router = useRouter();
+  const router = useRouterConnect();
   const { logoUrl } = useCompanyConfig();
   const [translate] = useTranslation();
   return (