+++ src/modules/auth/components/SignInTemplate/SignInTemplate.tsx
@@ -15,7 +15,7 @@ import { LocalStorageFields } from '../../../shared/enums/LocalStorageFields';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import { useCompanyConfig } from '../../../shared/hooks/useCompanyConfig';
 import { usePixwaySession } from '../../../shared/hooks/usePixwaySession';
-import useRouter from '../../../shared/hooks/useRouter';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
 import { useTimedBoolean } from '../../../shared/hooks/useTimedBoolean';
 import useTranslation from '../../../shared/hooks/useTranslation';
 import { usePasswordValidationSchema } from '../../hooks/usePasswordValidationSchema';
@@ -60,7 +60,7 @@ const _SignInTemplate = ({
   const { data: session } = usePixwaySession();
   const [isLoading, setIsLoading] = useState(false);
   const [isShowingErrorMessage, showErrorMessage] = useTimedBoolean(6000);
-  const router = useRouter();
+  const router = useRouterConnect();
   const { data: profile } = useProfile();
   const [callbackUrl, setCallbackUrl] = useLocalStorage<string>(
     LocalStorageFields.AUTHENTICATION_CALLBACK,
@@ -84,7 +84,7 @@ const _SignInTemplate = ({
   });
 
   useEffect(() => {
-    if (session) router.push(getRedirectUrl());
+    if (session) router.pushConnect(getRedirectUrl());
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [session, router]);
 
@@ -162,7 +162,9 @@ const _SignInTemplate = ({
                   error={fieldState.error}
                 />
                 <Link
-                  href={PixwayAppRoutes.REQUEST_PASSWORD_CHANGE}
+                  href={router.routerToHref(
+                    PixwayAppRoutes.REQUEST_PASSWORD_CHANGE
+                  )}
                   className="pw-text-[#383857] pw-text-[13px] pw-leading-[19.5px] hover:pw-underline hover:pw-text-[#5682C3] pw-underline"
                 >
                   {translate('auth>passwordChange>requestChangeFormTitle')}
@@ -185,7 +187,7 @@ const _SignInTemplate = ({
                 <Trans i18nKey={'auth>signIn>signUpCTA'}>
                   Não tem conta ainda?
                   <Link
-                    href={PixwayAppRoutes.SIGN_UP}
+                    href={router.routerToHref(PixwayAppRoutes.SIGN_UP)}
                     className="pw-text-brand-primary pw-underline"
                   >
                     Cadastre-se.