+++ src/modules/auth/components/VerifySignUpMailSentWithoutLayout/VerifySignUpMailSentWithoutLayout.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { Trans } from 'react-i18next';
 import { useLocalStorage } from 'react-use';
 
@@ -7,11 +7,11 @@ import { addMinutes, isAfter } from 'date-fns';
 import { LocalStorageFields } from '../../../shared/enums/LocalStorageFields';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import useCountdown from '../../../shared/hooks/useCountdown/useCountdown';
-import useRouter from '../../../shared/hooks/useRouter';
 import useTranslation from '../../../shared/hooks/useTranslation';
 import { ReactComponent as MailSent } from '../../assets/icons/mailSent.svg';
 import { useRequestPasswordChange } from '../../hooks';
 import { useEmailProtectedLabel } from '../../hooks/useEmailProtectedLabel';
+import { useRequestConfirmationMail } from '../../hooks/useRequestConfirmationMail';
 
 interface PasswordChangeMailSentProps {
   email: string;
@@ -20,26 +20,23 @@ interface PasswordChangeMailSentProps {
 }
 
 export const VerifySignUpMailSentWithoutLayout = ({
-  email = '',
+  email,
   isPostSignUp = false,
 }: PasswordChangeMailSentProps) => {
   const [translate] = useTranslation();
-  const { query, isReady } = useRouter();
-  const [emailToUse, setEmailToUse] = useState(email);
   const { mutate, isSuccess, isLoading, reset } = useRequestPasswordChange();
+  const {
+    mutate: emailMutate,
+    isSuccess: emailSuccess,
+    isLoading: emailLoading,
+    reset: emailReset,
+  } = useRequestConfirmationMail();
   const { minutes, seconds, setNewCountdown, isActive } = useCountdown();
   const [countdownDate, setCountdownDate] = useLocalStorage<Date>(
     LocalStorageFields.EMAIL_CONFIRMATION_LINK_COUNTDOWN_DATE
   );
 
-  useEffect(() => {
-    if (isReady && (!email || email === '') && query.email) {
-      setEmailToUse(query.email as string);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [query, isReady]);
-
-  const formattedEmail = useEmailProtectedLabel(emailToUse);
+  const formattedEmail = useEmailProtectedLabel(email);
 
   useEffect(() => {
     if (countdownDate && isAfter(new Date(countdownDate), new Date())) {
@@ -54,9 +51,31 @@ export const VerifySignUpMailSentWithoutLayout = ({
     }
   }, [isSuccess, setCountdownDate, reset]);
 
+  useEffect(() => {
+    if (emailSuccess) {
+      setCountdownDate(addMinutes(new Date(), 3));
+      emailReset();
+    }
+  }, [emailSuccess, setCountdownDate, emailReset]);
+
   const callbackPath = isPostSignUp
     ? PixwayAppRoutes.COMPLETE_SIGNUP
     : PixwayAppRoutes.SIGN_UP_MAIL_CONFIRMATION;
+
+  const handleClick = () => {
+    if (isPostSignUp) {
+      mutate({
+        email,
+        callbackPath,
+        verificationType: 'numeric',
+      });
+    } else {
+      emailMutate({
+        email,
+        callbackPath,
+      });
+    }
+  };
 
   return (
     <div className="pw-pt-0 pw-pb-6 sm:pw-mt-6 pw-flex pw-flex-col pw-items-center pw-leading-[23px] pw-text-lg">
@@ -72,15 +91,9 @@ export const VerifySignUpMailSentWithoutLayout = ({
       <div className="pw-mb-[23px]">
         <div className="pw-flex pw-justify-center">
           <button
-            disabled={isActive || isLoading}
+            disabled={isActive || isLoading || emailLoading}
             className="pw-font-semibold pw-text-[14px] pw-leading-[21px] pw-underline pw-text-brand-primary pw-font-poppins disabled:pw-text-[#676767] disabled:hover:pw-no-underline"
-            onClick={() =>
-              mutate({
-                email: emailToUse,
-                callbackPath,
-                verificationType: 'numeric',
-              })
-            }
+            onClick={handleClick}
           >
             {translate('auth>mailStep>resentCodeButton')}
           </button>
@@ -105,15 +118,9 @@ export const VerifySignUpMailSentWithoutLayout = ({
         <Trans i18nKey="auth>emailConfirmation>linkExpiresMessage">
           O link expira em 15 minutos
           <button
-            disabled={isActive || isLoading}
+            disabled={isActive || isLoading || emailLoading}
             className="pw-font-poppins pw-underline pw-font-semibold pw-leading-[19.5px] disabled:pw-text-[#676767]"
-            onClick={() =>
-              mutate({
-                email: emailToUse,
-                callbackPath,
-                verificationType: 'numeric',
-              })
-            }
+            onClick={handleClick}
           >
             Reenviar código
           </button>