+++ src/modules/auth/hooks/useSignUp/useSignUp.ts
@@ -1,13 +1,21 @@
 import { useMutation } from 'react-query';
 
-import { SignupUserDto } from '@w3block/sdk-id';
+import { I18NLocaleEnum } from '@w3block/sdk-id';
 
 import { PixwayAPIRoutes } from '../../../shared/enums/PixwayAPIRoutes';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import { useCompanyConfig } from '../../../shared/hooks/useCompanyConfig';
 import { useGetW3blockIdSDK } from '../../../shared/hooks/useGetW3blockIdSDK';
 
-type Payload = Pick<SignupUserDto, 'confirmation' | 'email' | 'password'>;
+interface Payload {
+  password: string;
+  confirmation: string;
+  email: string;
+  tenantId?: string;
+  name?: string;
+  i18nLocale?: I18NLocaleEnum;
+  callbackUrl?: string;
+}
 
 export const useSignUp = () => {
   const getSDK = useGetW3blockIdSDK();
@@ -18,7 +26,7 @@ export const useSignUp = () => {
       ...payload,
       tenantId: companyId,
       callbackUrl: new URL(
-        PixwayAppRoutes.SIGN_UP_MAIL_CONFIRMATION,
+        payload.callbackUrl ?? PixwayAppRoutes.SIGN_UP_MAIL_CONFIRMATION,
         appBaseUrl
       ).toString(),
     });