+++ src/modules/auth/templates/CompleteProfileCustomTemplate/CompleteProfileCustomTemplate.tsx
@@ -9,13 +9,13 @@ import TranslatableComponent from '../../../shared/components/TranslatableCompon
 import { FAQContextEnum } from '../../../shared/enums/FAQContext';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import { useCompanyConfig } from '../../../shared/hooks/useCompanyConfig';
-import useRouter from '../../../shared/hooks/useRouter';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
 import useTranslation from '../../../shared/hooks/useTranslation';
 import { AuthLayoutBaseClasses } from '../../components/AuthLayoutBase';
 import { SignUpFormData } from '../../components/SignUpForm/interface';
 import { useChangePassword } from '../../hooks/useChangePassword';
 import { usePixwayAuthentication } from '../../hooks/usePixwayAuthentication';
-import { CompleteSigUpSuccessTemplateSDK } from '../CompleteSignUpSuccessTemplateSDK';
+import { CompleteSignUpSuccessTemplateSDK } from '../CompleteSignUpSuccessTemplateSDK';
 import { SignUpTemplateSDK } from '../SignUpTemplateSDK/SignUpTemplateSDK';
 import { VerifySignUpMailSentTemplateSDK } from '../VerifySignUpMailSentTemplateSDK';
 import { VerifySignUpTokenExpiredTemplateSDK } from '../VerifySignUpTokenExpiredTemplateSDK';
@@ -59,7 +59,7 @@ const _CompleteProfileCustomTemplate = ({
   verifyEmailProps,
   extraBy,
 }: CompleteProfileCustomTemplateProps) => {
-  const router = useRouter();
+  const router = useRouterConnect();
   const [translate] = useTranslation();
   const { companyId } = useCompanyConfig();
   const { email, token } = router.query;
@@ -87,6 +87,9 @@ const _CompleteProfileCustomTemplate = ({
     if (token && getIsTokenExpired(token as string)) {
       setStep(Steps.TOKEN_EXPIRED);
     }
+    if (token && !getIsTokenExpired(token as string)) {
+      setStep(Steps.MAIL_CONFIRMED);
+    }
   }, [token]);
 
   const onSubmit = ({ confirmation, password }: SignUpFormData) => {
@@ -104,7 +107,9 @@ const _CompleteProfileCustomTemplate = ({
             password,
             companyId,
           })
-            .then(() => router.push(PixwayAppRoutes.CONNECT_EXTERNAL_WALLET))
+            .then(() =>
+              router.pushConnect(PixwayAppRoutes.CONNECT_EXTERNAL_WALLET)
+            )
             .catch((e) => {
               // eslint-disable-next-line no-console
               console.log(e);
@@ -147,11 +152,7 @@ const _CompleteProfileCustomTemplate = ({
       }
     />
   ) : (
-    <CompleteSigUpSuccessTemplateSDK
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      props={verifyEmailProps!}
-      extraBy={extraBy}
-    />
+    <CompleteSignUpSuccessTemplateSDK {...verifyEmailProps} extraBy={extraBy} />
   );
 };