+++ src/modules/shared/components/MyProfile/MyProfile.tsx
@@ -12,6 +12,7 @@ import { ReactComponent as EyeCrossedIcon } from '../../../shared/assets/icons/e
 import { ReactComponent as MetamaskIcon } from '../../../shared/assets/icons/metamask.svg';
 import { ReactComponent as WalletIcon } from '../../../shared/assets/icons/walletOutlined.svg';
 import { PixwayAppRoutes } from '../../enums/PixwayAppRoutes';
+import { useRouterConnect } from '../../hooks';
 import { usePatchProfile, useProfile } from '../../hooks/useProfile';
 import useTranslation from '../../hooks/useTranslation';
 import { useUserWallet } from '../../hooks/useUserWallet';
@@ -22,6 +23,7 @@ import { Spinner } from '../Spinner';
 
 export const MyProfile = () => {
   const { data: profile } = useProfile();
+  const router = useRouterConnect();
   const [showValue, toggleShowValue] = useToggle(false);
   const [nameVal, setNameVal] = useState('');
   const [translate] = useTranslation();
@@ -78,7 +80,7 @@ export const MyProfile = () => {
               }
             )
           )}
-          <Link href={PixwayAppRoutes.ADD_FUNDS_TYPE}>
+          <Link href={router.routerToHref(PixwayAppRoutes.ADD_FUNDS_TYPE)}>
             <div className="pw-w-[165px] pw-bg-brand-primary pw-p-[8px_16px_8px_11px] pw-border-2 pw-border-[#353945] pw-rounded-[48px] pw-flex pw-justify-start pw-items-center pw-gap-2">
               <div className="pw-rounded-full pw-border pw-bg-brand-primary pw-border-white pw-w-[30px] pw-h-[30px] pw-p-[5px] pw-flex pw-justify-center pw-items-center">
                 <CashIcon className="pw-fill-white" />