+++ src/modules/shared/hooks/useGuardedPage/useGuardedPage.tsx
@@ -4,19 +4,19 @@ import { usePixwayAuthentication } from '../../../auth/hooks/usePixwayAuthentica
 import { validateJwtToken } from '../../config/api';
 import { PixwayAppRoutes } from '../../enums/PixwayAppRoutes';
 import { usePixwaySession } from '../usePixwaySession';
-import useRouter from '../useRouter';
+import { useRouterConnect } from '../useRouterConnect';
 
 export const useGuardedPage = (routeToReturn = PixwayAppRoutes.SIGN_IN) => {
   const { data: session } = usePixwaySession();
   const { signOut } = usePixwayAuthentication();
-  const { push, isReady } = useRouter();
+  const { isReady, pushConnect } = useRouterConnect();
   useEffect(() => {
     if (
       isReady &&
       (!session || !validateJwtToken(session.accessToken as string))
     ) {
       signOut().then(() => {
-        push(routeToReturn);
+        pushConnect(routeToReturn);
       });
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps