+++ src/modules/tokens/components/TokenDetailsTemplate/TokenDetailsTemplate.tsx
@@ -3,7 +3,7 @@ import { ReactComponent as ChevronLeft } from '../../../shared/assets/icons/chev
 import { MintedInfoCard } from '../../../shared/components/MintedInfoCard';
 import TranslatableComponent from '../../../shared/components/TranslatableComponent';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
-import useRouter from '../../../shared/hooks/useRouter';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
 import { usePublicTokenData } from '../../hooks/usePublicTokenData';
 import {
   Dimensions2DValue,
@@ -12,7 +12,7 @@ import {
 import { TokenDetailsCard } from '../TokenDetailsCard';
 
 const _TokenDetailsTemplate = () => {
-  const router = useRouter();
+  const router = useRouterConnect();
   const contractAddress = (router.query.contractAddress as string) ?? '';
   const chainId = (router.query.chainId as string) ?? '';
   const tokenId = (router.query.tokenId as string) ?? '';
@@ -27,7 +27,7 @@ const _TokenDetailsTemplate = () => {
       <div className="sm:pw-hidden pw-mb-4">
         <div className="pw-flex pw-items-center pw-gap-x-4 pw-mb-[16.5px]">
           <button
-            onClick={() => router.push(PixwayAppRoutes.TOKENS)}
+            onClick={() => router.pushConnect(PixwayAppRoutes.TOKENS)}
             className="pw-flex pw-items-center pw-justify-center pw-rounded-full pw-border pw-border-[#777E8F] pw-bg-[#F7F7F7] pw-w-4 pw-h-4 pw-ml-[22px]"
           >
             <ChevronLeft className="pw-w-[4.25px] pw-h-[8.5px] pw-fill-[#777E8F]" />